pyproject.toml
[tool.poetry]
name = "lms-fastapi"
version = "0.1.0"
description = ""
authors = ["therollingambit <therollingambit@gmail.com>"]
readme = "README.md"
packages = [{include = "lms_fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.21.1"
fastapi = "^0.95.0"
pydantic = "^1.10.7"
sqlalchemy = "^2.0.9"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.40.0"
alembic = "^1.10.3"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
